+++ CHANGELOG.md
@@ -37,6 +37,11 @@ A more detailed list of changes is available in the corresponding milestones for
   - **[com.google.fonts/check/glyf_unused_data]:** Added rationale. (PR #4570)
   - **[com.google.fonts/check/points_out_of_bounds]:** Added rationale. (PR #4570)
   - **[com.google.fonts/check/gpos_kerning_info]:** Added rationale. (PR #4570)
+  - **[com.google.fonts/check/maxadvancewidth]:** Added rationale. (PR #4570)
+  - **[com.google.fonts/check/loca/maxp_num_glyphs]:** Added rationale. (PR #4570)
+  - **[com.google.fonts/check/name/no_copyright_on_description]:** Added rationale. (PR #4570)
+  - **[com.adobe.fonts/check/postscript_name]:** Added rationale. (PR #4570)
+  - **[com.google.fonts/check/family_naming_recommendations]:** Added rationale. (PR #4570)
 
 
 ## 0.12.0a2 (2024-Feb-21)